Coding: rename file
Filename should match name of main class, DymoLabeler
